Decouple WhatsAppAdapter outbound by BSP
- DRY out the code by separating the two implementations of WhatsApp by
  business solution provider
